refactor(telemetry): simplify code layout
Problem:
Too many files, too granular, too spread out.

Solution:
- Merge various telemetry files.
- Use ClassToInterfaceType.
