+++ src/test/globalSetup.test.ts
@@ -14,7 +14,7 @@ import { CodelensRootRegistry } from '../shared/sam/codelensRootRegistry'
 import { CloudFormationTemplateRegistry } from '../shared/cloudformation/templateRegistry'
 import { getLogger, LogLevel } from '../shared/logger'
 import { setLogger } from '../shared/logger/logger'
-import { DefaultTelemetryService } from '../shared/telemetry/defaultTelemetryService'
+import { DefaultTelemetryService } from '../shared/telemetry/telemetryService'
 import * as fakeTelemetry from './fake/fakeTelemetryService'
 import { TestLogger } from './testLogger'
 import { FakeAwsContext } from './utilities/fakeAwsContext'
